this commit will extract grocery list fetch and isAuthenticated into separate files
