+++ backend/routes/groceryListRoutes.ts
@@ -13,21 +13,7 @@ const { PrismaClient } = require("@prisma/client");
 const prisma = new PrismaClient();
 
 import { unitConversions } from "../utils/constants";
-
-// TODO move into utils file to prevent code reuse
-// middleware to check if a user is authenticated
-const isAuthenticated = (
-  req: includeSession,
-  res: Response,
-  next: NextFunction
-) => {
-  if (!req.session.userId) {
-    return res
-      .status(401)
-      .json({ error: "You must be logged in to perform this action." });
-  }
-  next();
-};
+import { isAuthenticated } from "../utils/authMiddleware";
 
 router.get("/", isAuthenticated, async (req: Request, res: Response) => {
   const userId = req.session.userId;
