+++ backend/routes/ingredientsRoutes.ts
@@ -6,20 +6,7 @@ const router = express.Router();
 
 const { PrismaClient } = require("@prisma/client");
 const prisma = new PrismaClient();
-
-// middleware to check if a user is authenticated
-const isAuthenticated = (
-  req: includeSession,
-  res: Response,
-  next: NextFunction
-) => {
-  if (!req.session.userId) {
-    return res
-      .status(401)
-      .json({ error: "You must be logged in to perform this action." });
-  }
-  next();
-};
+import { isAuthenticated } from "../utils/authMiddleware";
 
 type IngredientOnHand = {
   expirationDate: string;
