+++ backend/utils/authMiddleware.ts
@@ -1,0 +1,18 @@
+import { includeSession } from "../types/session";
+import { Response, NextFunction } from "express";
+
+// middleware to check if a user is authenticated
+const isAuthenticated = (
+  req: includeSession,
+  res: Response,
+  next: NextFunction
+) => {
+  if (!req.session.userId) {
+    return res
+      .status(401)
+      .json({ error: "You must be logged in to perform this action." });
+  }
+  next();
+};
+
+export { isAuthenticated };
